+++ stdnet/orm/base.py
@@ -238,6 +238,14 @@ the model table'''
             raise errorClass('Cannot Order by attribute "{0}".\
  It is not a scalar field.'.format(sortby))
 
+    def multifields_ids_todelete(self, instance):
+        '''Return the list of ids of multifields belonging to *instance*
+ which needs to be deleted when *instance* is deleted.'''
+        gen = (field.id(instance) for field in self.multifields\
+                                         if field.todelete())
+        return [fid for fid in gen if fid]
+    
+    
 class FakeMeta(object):
     pass